add friendly message for tcurl health ping
