+++ log.js
@@ -1,0 +1,54 @@
+#!/usr/bin/env node
+
+// Copyright (c) 2015 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+'use strict';
+
+var console = require('console');
+var util = require('util');
+module.exports.display = display;
+module.exports.setup = setup;
+
+var options = null;
+
+function setup(opts) {
+    options = opts;
+}
+
+function display(level, value) {
+    if (options.json) {
+        log(level, JSON.stringify(value, null, options.json));
+    } else if (options.raw) {
+        log(level, String(value));
+    } else {
+        log(level, util.inspect(value, {
+            depth: options.depth || 2
+        }));
+    }
+}
+
+function log(level, message) {
+    if (level === 'error') {
+        console.error(message);
+    } else {
+        console.log(message);
+    }
+}
